third_party/WebKit/LayoutTests/html/dialog/dialog-autofocus-multiple-times-expected.txt

Tests autofocus is run every time a dialog is opened.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS document.activeElement is document.getElementById("outer-button")
PASS document.activeElement is document.getElementById("input2")
PASS document.activeElement is document.getElementById("input1")
PASS document.activeElement is document.getElementById("input2")
PASS document.activeElement is document.getElementById("input1")
PASS document.activeElement is document.getElementById("input2")
PASS document.activeElement is document.getElementById("input1")
PASS 3 is 3
PASS successfullyParsed is true

TEST COMPLETE
